+++ src/rendering/shaders/default.vert.glsl
@@ -1,11 +1,10 @@
 #version 300 es
-precision mediump float;                        // Set default float precision
+precision highp float;                        // Set default float precision
 
 in vec4 pos, col, uv, normal;                 // Vertex attributes: position, color, texture coordinates, normal (if any)
 uniform mat4 pv, eye, m, im;                  // Uniform transformation matrices: projection * view, eye, model, inverse model
 uniform vec4 bb;                              // If the current shape is a billboard: bb = [w, h, 1.0, 0.0]
 out vec4 v_pos, v_col, v_uv, v_normal;  
-
 
 // Varyings sent to the fragment shader: position, color, texture coordinates, normal (if any)
 void main() {
@@ -16,5 +15,5 @@ void main() {
     );
     v_col = col;                                // Set varyings 
     v_uv = uv;
-    v_normal = transpose(inverse(m)) * normal;  // recompute normals to match model thansformation
+    v_normal = transpose(inverse(m)) * normal;  // recompute normals to match model transformation
 }
